Refatoração do cadastro de usuário
